requirements.txt
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pytesseract==0.3.13
requests==2.32.3
